Added about section
Added the about section underneath the home page (although the order of the website may change later on). A placeholder for the about text has been added that can be adjusted later, as well as an image that will be replaced with something more appropriate when the details of the portfolio are being finalised.