Created 3 web pages for each project
Created a web page for each project in the portfolio. Each one shares the same layout so it remains consistent, and the header bar from the main screen is also at the top of each page. The text and images will need to be updated on each one, and a short video will also need to be created to go into the landing section of each page.